Apply name normalization to competition detail page
Extracts normalizeName into a shared utility and applies it to all
four name render sites in CompetitionPage.js.
